Reworked the Node class to use Node instead of File type for parent fields
Instead of passing a File object for the parent and storing it like this, I reworked the class in order to allow the parent to be stored as a Node object instead. Storing a File object defeats the purpose of creating the Node class in the first place. I also added the getParentPath() method that essentially provides the same functionality that the getParent() method used to serve.