Added duplicate file method, and more
Added several helper methods and a findDuplicateFiles() method that will be used to find duplicate files in a FileTree. Right now there is not yet a way to call it from the main program, that has not been implemented yet. But, that will be added soon after more testing is done. Read documentation above each method for more information.